More accurate
Now it should be as accurate as you said! The current code is terribly redundant 😔 We should definitely use **uisleep** because **sleep** is based on the users FPS. So if the mission struggles loading or lacks for a while at the beginning (which happens most often to players with lower specs) than the soundtrack continues but does not match the sleep time and then aborts at some point.